+++ uefi/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `proto::shell::Shell::{var(), set_var(), vars()}`
 - Added `proto::pci::root_bridge::PciRootBridgeIo::configuration()`.
 - Added `proto::pci::root_bridge::PciRootBridgeIo::enumerate()`.
+- Added `proto::nvme::pass_thru::NvmePassThru::broadcast()`.
 
 ## Changed
 - Changed ordering of `proto::pci::PciIoAddress` to (bus -> dev -> fun -> reg -> ext_reg).